requirements.txt
torch
torchvision
ultralytics
streamlit
pandas
scikit-learn
pillow